+++ components/esp_matter/esp_matter_attribute.cpp
@@ -14,6 +14,8 @@
 
 #include <esp_log.h>
 #include <esp_matter_attribute.h>
+#include <esp_matter.h>
+#include <esp_matter_core.h>
 
 static const char *TAG = "esp_matter_attribute";
 
@@ -173,7 +175,7 @@ attribute_t *create_node_label(cluster_t *cluster, char *value, uint16_t length)
 {
     return esp_matter::attribute::create(cluster, BasicInformation::Attributes::NodeLabel::Id,
                                          ATTRIBUTE_FLAG_WRITABLE | ATTRIBUTE_FLAG_NONVOLATILE,
-                                         esp_matter_char_str(value, length));
+                                         esp_matter_char_str(value, length), k_max_node_label_length);
 }
 
 attribute_t *create_location(cluster_t *cluster, char *value, uint16_t length)
@@ -719,7 +721,7 @@ attribute_t *create_node_label(cluster_t *cluster, char *value, uint16_t length)
 {
     return esp_matter::attribute::create(cluster, BridgedDeviceBasicInformation::Attributes::NodeLabel::Id,
                                          ATTRIBUTE_FLAG_NONVOLATILE | ATTRIBUTE_FLAG_WRITABLE,
-                                         esp_matter_char_str(value, length));
+                                         esp_matter_char_str(value, length), k_max_node_label_length);
 }
 
 attribute_t *create_reachable(cluster_t *cluster, bool value)
@@ -1867,9 +1869,13 @@ namespace attribute {
 
 attribute_t *create_active_locale(cluster_t *cluster, char *value, uint16_t length)
 {
+    if (length > k_max_active_locale_length) {
+        ESP_LOGE(TAG, "Could not create attribute, string length out of bound");
+        return NULL;
+    }
     return esp_matter::attribute::create(cluster, LocalizationConfiguration::Attributes::ActiveLocale::Id,
                                          ATTRIBUTE_FLAG_WRITABLE | ATTRIBUTE_FLAG_NONVOLATILE,
-                                         esp_matter_char_str(value, length));
+                                         esp_matter_char_str(value, length), k_max_active_locale_length);
 }
 
 attribute_t *create_supported_locales(cluster_t *cluster, uint8_t *value, uint16_t length, uint16_t count)